+++ cmd/docgen/po/testdata/locale/es/simple.po
